requirements.txt
fastapi
uvicorn
pydantic
sqlite3  # Comes with Python by default
pandas
plotly
google-generativeai